Jeszcze bez komentarzy, do tyg 1

+++ inc/LZespolona.hh
@@ -27,7 +27,7 @@ LZespolona WczytajLicz();
 
 LZespolona Sprzezenie(LZespolona Z1);
 
-float Modul(LZespolona Z1);
+double Modul(LZespolona Z1);
 
 LZespolona operator + (LZespolona  Skl1,  LZespolona  Skl2);
 
@@ -35,7 +35,7 @@ LZespolona operator - (LZespolona Odjemna, LZespolona Odjemnik);
 
 LZespolona operator * (LZespolona Czyn1, LZespolona Czyn2);
 
-LZespolona operator / (LZespolona Z1, float Dzielnik);
+LZespolona operator / (LZespolona Z1, double Dzielnik);
 LZespolona operator / (LZespolona Dzielna, LZespolona Dzielnik);
 
 #endif
